Inline TypedPValue
This has exactly one implementation, and this is not expected to change.
